# backend/src/main/resources/static/swagger-api/swagger.yaml
openapi: 3.0.0
info:
  version: "1.0.0"
  title: "ecommerce-spring-reactjs"
servers:
  - url: "http://localhost:8080"
...
paths:
...
  /api/v1/registration:
    post:
      tags:
        - "RegistrationController"
      summary: "Registration"
      operationId: "registration"
      requestBody:
        required: true
        content:
          'application/json':
            schema:
              type: object
              properties:
                email:
                  type: "string"
                  example: "test1234@test.com"
                password:
                  type: "string"
                  example: "1234567890"
                password2:
                  type: "string"
                  example: "1234567890"
                firstName:
                  type: "string"
                  example: "John"
                lastName:
                  type: "string"
                  example: "Doe"
                captcha:
                  type: "string"
      responses:
        "200":
          description: "OK"
          content:
            'application/json':
              schema:
                type: object
                properties:
                  data:
                    type: "string"
                    example: "User successfully registered."
        "400":
          description: "Bad Request"
          content:
            'application/json':
              schema:
                type: object
                properties:
                  passwordError:
                    type: "string"
                  password2Error:
                    type: "string"
              examples:
                PasswordMatchError:
                  summary: "Password Match Error"
                  value:
                    passwordError: "Passwords do not match."
                EmailError:
                  summary: "Email Error"
                  value:
                    emailError: "Email is already used."
                BindingResultError:
                  summary: "Binding Result Errors"
                  value:
                    captchaError: "Fill captcha."
                    firstNameError: "First name cannot be empty"
                    lastNameError: "Last name cannot be empty"
                    passwordError: "The password must be between 6 and 16 characters long"
                    password2Error: "The password confirmation must be between 6 and 16 characters long"
                    emailError: "Incorrect email"
  /api/v1/registration/activate/{code}:
    get:
      tags:
        - "RegistrationController"
      summary: "Activate email code"
      operationId: "activateEmailCode"
      parameters:
        - name: "code"
          in: "path"
          required: true
          schema:
            type: "string"
      responses:
        "200":
          description: "OK"
          content:
            'application/json':
              schema:
                type: object
                properties:
                  data:
                    type: "string"
                    example: "User successfully activated."
        "404":
          description: "Bad Request"
          content:
            'application/json':
              schema:
                type: object
                properties:
                  data:
                    type: string
                    example: "Activation code not found."
  /api/v1/users/info:
    get:
      tags:
        - "UserController"
      summary: "Get authenticated user info"
      operationId: "getUserInfo"
      parameters: []
      responses:
        "200":
          description: "OK"
          content:
            'application/json':
              schema:
                $ref: "#/components/schemas/User"
      security:
        - token: []
  /api/v1/users/graphql/info:
    post:
      tags:
        - "UserController"
      summary: "Get authenticated user info by GraphQL query"
      operationId: "getUserInfoByQuery"
      requestBody:
        required: true
        content:
          'application/json':
            schema:
              type: object
              properties:
                query:
                  type: "string"
                  example: "{ user(id: 2) { id email password firstName lastName city address phoneNumber postIndex activationCode passwordResetCode active provider roles } }"
      responses:
        "200":
          description: "OK"
          content:
            'application/json':
              schema:
                type: object
                properties:
                  id:
                    type: "integer"
                    format: int64
                  email:
                    type: "string"
                  firstName:
                    type: "string"
                  lastName:
                    type: "string"
                  city:
                    type: "string"
                  address:
                    type: "string"
                  phoneNumber:
                    type: "string"
                  postIndex:
                    type: "string"
                example:
                  id: 2
                  email: "test123@test.com"
                  firstName: "John"
                  lastName: "Doe"
                  city: "New York"
                  address: "Wall Street 1"
                  phoneNumber: "1234567890"
                  postIndex: "1234567890"
      security:
        - token: []
  /api/v1/users/edit:
    put:
      tags:
        - "UserController"
      summary: "Edit authenticated user info"
      operationId: "updateUserInfo"
      requestBody:
        content:
          'application/json':
            schema:
              type: object
              properties:
                firstName:
                  type: "string"
                  example: "John"
                lastName:
                  type: "string"
                  example: "Doe"
                city:
                  type: "string"
                  example: "New York"
                address:
                  type: "string"
                  example: "Wall Street 1"
                phoneNumber:
                  type: "string"
                  example: "123456"
                postIndex:
                  type: "string"
                  example: "123456"
              required:
                - firstName
                - lastName
      responses:
        "200":
          description: "OK"
          content:
            'application/json':
              schema:
                $ref: "#/components/schemas/User"
        "400":
          description: "Bad Request"
          content:
            'application/json':
              schema:
                type: object
                properties:
                  firstNameError:
                    type: string
                    example: "First name cannot be empty"
                  lastNameError:
                    type: string
                    example: "Last name cannot be empty"
      security:
        - token: []
...
  /api/v1/users/cart:
    post:
      tags:
        - "UserController"
      summary: "Get perfumes by ids"
      operationId: "getCart"
      requestBody:
        content:
          'application/json':
            schema:
              type: "array"
              items:
                type: "integer"
                format: int32
              example: [1, 2, 3]
      responses:
        "200":
          description: "OK"
          content:
            'application/json':
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/PerfumeResponse"
  /api/v1/users/orders:
    get:
      tags:
        - "UserController"
      summary: "Get all user orders by email"
      operationId: "getAllUserOrders"
      parameters: []
      responses:
        "200":
          description: "OK"
          content:
            'application/json':
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/OrderResponse"
      security:
        - token: []
  /api/v1/user/graphql/orders:
    post:
      tags:
        - "UserController"
      summary: "Get all user orders by GraphQL query"
      operationId: "getUserOrdersByQuery"
      requestBody:
        required: true
        content:
          'application/json':
            schema:
              type: object
              properties:
                query:
                  type: "string"
                  example: "{ ordersByEmail(email: test123@test.com) { id totalPrice date firstName lastName city address email phoneNumber postIndex orderItems { id amount quantity perfume { id perfumeTitle perfumer price filename } } } }"
      responses:
        "200":
          description: "OK"
          content:
            'application/json':
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/OrderResponse"
      security:
        - token: []
  /api/v1/users/order:
    post:
      tags:
        - "UserController"
      summary: "Post order"
      operationId: "postOrder"
      requestBody:
        content:
          'application/json':
            schema:
              type: object
              properties:
                totalPrice:
                  type: "number"
                  format: "double"
                  example: 100
                firstName:
                  type: "string"
                  example: "John"
                lastName:
                  type: "string"
                  example: "Doe"
                city:
                  type: "string"
                  example: "New York"
                address:
                  type: "string"
                  example: "Wall Street 1"
                email:
                  type: "string"
                  example: "test123@test.com"
                phoneNumber:
                  type: "string"
                  example: "123456"
                postIndex:
                  type: "integer"
                  format: "int32"
                  example: 123456
                perfumesId:
                  type: "object"
                  properties:
                    43:
                      type: integer
                      example: 1
              required:
                - firstName
                - lastName
                - city
                - address
                - email
                - phoneNumber
                - postIndex
      responses:
        "200":
          description: "OK"
          content:
            'application/json':
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/OrderResponse"
        "400":
          description: "Bad Request"
          content:
            'application/json':
              schema:
                type: object
                properties:
                  firstNameError:
                    type: "string"
                    example: "Fill in the input field"
                  lastNameError:
                    type: "string"
                    example: "Fill in the input field"
                  cityError:
                    type: "string"
                    example: "Fill in the input field"
                  addressError:
                    type: "string"
                    example: "Fill in the input field"
                  emailError:
                    type: "string"
                    example: "Email cannot be empty"
                  phoneNumberError:
                    type: "string"
                    example: "Phone number cannot be empty"
                  postIndexError:
                    type: "string"
                    example: "Post index cannot be empty"
  /api/v1/users/review:
    post:
      tags:
        - "UserController"
      summary: "Add review to perfume"
      operationId: "addReviewToPerfume"
      requestBody:
        content:
          'application/json':
            schema:
              type: object
              properties:
                perfumeId:
                  type: "number"
                  format: "int64"
                  example: 43
                author:
                  type: "string"
                  example: "John Doe"
                message:
                  type: "string"
                  example: "Hello World!"
              required:
                - perfumeId
                - author
                - message
      responses:
        "200":
          description: "OK"
          content:
            'application/json':
              schema:
                type: object
                properties:
                  data:
                    type: string
                    example: "Review added successfully."
        "400":
          description: "Bad Request"
          content:
            'application/json':
              schema:
                type: object
                properties:
                  authorError:
                    type: "string"
                    example: "Fill in the input field"
                  messageError:
                    type: "string"
                    example: "Fill in the input field"
components:
  securitySchemes:
    token:
      type: apiKey
      name: Authorization
      in: header
  schemas:
    PerfumeResponse:
      type: "object"
      properties:
        id:
          type: "integer"
          format: "int64"
          example: 43
        perfumeTitle:
          type: "string"
          example: "Sauvage"
        perfumer:
          type: "string"
          example: "Dior"
        year:
          type: "integer"
          format: "int32"
          example: 2015
        country:
          type: "string"
          example: "France"
        perfumeGender:
          type: "string"
          example: "male"
        fragranceTopNotes:
          type: "string"
          example: "Sichuan Pepper"
        fragranceMiddleNotes:
          type: "string"
          example: "Bergamot"
        fragranceBaseNotes:
          type: "string"
          example: "Ambroxan"
        description:
          type: "string"
          example: "Perfume description"
        filename:
          type: "string"
          example: "Dior_Sauvage.jpg"
        price:
          type: "integer"
          format: "int32"
          example: 62
        volume:
          type: "string"
          example: "100"
        type:
          type: "string"
          example: "Eau de parfum"
        reviews:
          type: "array"
          items:
            $ref: "#/components/schemas/ReviewResponse"
        file:
          type: "string"
          format: binary
    ReviewResponse:
      type: "object"
      properties:
        id:
          type: "integer"
          format: "int64"
          example: 1
        author:
          type: "string"
          example: "John Doe"
        message:
          type: "string"
          example: "Message"
        date:
          type: "string"
          example: "2021-04-07"
...
    OrderResponse:
      type: "object"
      properties:
        id:
          type: "integer"
          format: "int64"
          example: 1
        totalPrice:
          type: "number"
          format: "double"
          example: 100
        date:
          type: "string"
          example: "2021-04-07"
        firstName:
          type: "string"
          example: "John"
        lastName:
          type: "string"
          example: "Doe"
        city:
          type: "string"
          example: "New York"
        address:
          type: "string"
          example: "Wall Street 1"
        email:
          type: "string"
          example: "test123@test.com"
        phoneNumber:
          type: "string"
          example: "123456"
        postIndex:
          type: "integer"
          format: "int32"
          example: 123456
        orderItems:
          type: "array"
          items:
            $ref: "#/components/schemas/OrderItemResponse"
    OrderItemResponse:
      type: "object"
      properties:
        id:
          type: "integer"
          format: "int64"
          example: 1
        amount:
          type: "integer"
          format: "int64"
          example: 100
        quantity:
          type: "integer"
          format: "int64"
          example: 2
        perfume:
          $ref: "#/components/schemas/PerfumeResponse"
    User:
      type: "object"
      properties:
        id:
          type: "integer"
          format: "int64"
          example: 2
        email:
          type: "string"
          example: "test123@test.com"
        firstName:
          type: "string"
          example: "John"
        lastName:
          type: "string"
          example: "Doe"
        city:
          type: "string"
          example: "New York"
        address:
          type: "string"
          example: "Wall Street 1"
        phoneNumber:
          type: "string"
          example: "123456"
        postIndex:
          type: "string"
          example: "123456"
        provider:
          type: "string"
          example: "LOCAL"
        active:
          type: "boolean"
          example: true
        activationCode:
          type: "string"
          example: ""
        passwordResetCode:
          type: "string"
          example: ""
        roles:
          type: "array"
          items:
            type: "string"
            example: "USER"
...